requirements.txt
flask
flask-cors
python-dotenv
supabase
requests
bcrypt
psycopg2-binary
